burn-core: Use the same padding structure for many operators
