+++ burn-core/src/nn/mod.rs
@@ -22,6 +22,7 @@ mod gelu;
 mod initializer;
 mod linear;
 mod norm;
+mod padding;
 mod pos_encoding;
 mod relu;
 mod rnn;
@@ -32,6 +33,7 @@ pub use gelu::*;
 pub use initializer::*;
 pub use linear::*;
 pub use norm::*;
+pub use padding::*;
 pub use pos_encoding::*;
 pub use relu::*;
 pub use rnn::*;
