+++ burn-core/src/nn/padding.rs
@@ -1,0 +1,41 @@
+use crate as burn;
+
+use burn_tensor::ops::conv::calculate_conv_padding;
+
+use crate::config::Config;
+use crate::module::Module;
+
+/// Padding configuration for many operators.
+#[derive(Module, Config, Debug)]
+pub enum PaddingConfig {
+    /// Dynamically calculate the amount of padding necessary to ensure that the output size will be
+    /// the same as the input.
+    Same,
+    /// Same as no padding.
+    Valid,
+    /// Applies the specified amount of padding to all inputs.
+    Explicit(usize, usize),
+}
+
+impl PaddingConfig {
+    pub(crate) fn calculate_padding_2d(
+        &self,
+        height: usize,
+        width: usize,
+        kernel_size: &[usize; 2],
+        stride: &[usize; 2],
+    ) -> [usize; 2] {
+        let same_padding = || {
+            let p1 = calculate_conv_padding(kernel_size[0], stride[0], height, height);
+            let p2 = calculate_conv_padding(kernel_size[1], stride[1], width, width);
+
+            [p1, p2]
+        };
+
+        match self {
+            Self::Same => same_padding(),
+            Self::Valid => [0, 0],
+            Self::Explicit(v1, v2) => [*v1, *v2],
+        }
+    }
+}
