+++ burn-core/src/nn/pool/max_pool2d.rs
@@ -2,7 +2,7 @@ use crate as burn;
 
 use crate::config::Config;
 use crate::module::Module;
-use crate::nn::conv::Conv2dPaddingConfig;
+use crate::nn::PaddingConfig;
 use crate::tensor::backend::Backend;
 use crate::tensor::Tensor;
 use burn_tensor::module::max_pool2d;
@@ -18,19 +18,16 @@ pub struct MaxPool2dConfig {
     #[config(default = "[1, 1]")]
     pub strides: [usize; 2],
     /// The padding configuration.
-    #[config(default = "MaxPool2dPaddingConfig::Valid")]
-    pub padding: MaxPool2dPaddingConfig,
+    #[config(default = "PaddingConfig::Valid")]
+    pub padding: PaddingConfig,
 }
-
-/// Padding configuration for 2D max pooling [config](MaxPool2dConfig).
-pub type MaxPool2dPaddingConfig = Conv2dPaddingConfig;
 
 /// Applies a 2D max pooling over input tensors.
 #[derive(Module, Debug, Clone)]
 pub struct MaxPool2d {
     stride: [usize; 2],
     kernel_size: [usize; 2],
-    padding: MaxPool2dPaddingConfig,
+    padding: PaddingConfig,
 }
 
 impl MaxPool2dConfig {
